requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
apscheduler>=3.10.0
structlog>=23.0.0
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio
